+++ prefs.js
@@ -242,6 +242,7 @@ export default class AstraMonitorPrefs extends ExtensionPreferences {
             subtitle:  this.tab + _('Set icon name (ie: \'cpu-symbolic\')') + '\n' + this.tab + _('Set to empty to disable icon override'),
         }, 'processor-header-icon-custom', iconSection, '');
         this.addColorRow({ title: this.tab + _('Icon Color') }, 'processor-header-icon-color', iconSection, '');
+        this.addColorRow({ title: this.tab + _('Icon Alert Color') }, 'processor-header-icon-alert-color', iconSection, 'rgba(235, 64, 52, 255)');
         this.addSpinRow({
             title: this.tab + _('Icon Size'),
             subtitle:  this.tab + _('Experimental feature: may require to disable/enable the extension.'),
@@ -253,7 +254,11 @@ export default class AstraMonitorPrefs extends ExtensionPreferences {
         
         const percentageSection = this.addExpanderRow(_('Percentage'), group);
         this.addSwitchRow(this.tab + _('Show Percentage'), 'processor-header-percentage', percentageSection);
-        this.addSwitchRow(this.tab + _('Show Percentage Single Core'), 'processor-header-percentage-core', percentageSection);
+        this.addSwitchRow(this.tab + _('Single Core Percentage'), 'processor-header-percentage-core', percentageSection);
+        this.addSpinRow({
+            title: this.tab + _('Icon Alert'),
+            subtitle:  this.tab + _('Set 0 to disable. Value is % of total cpu.'),
+        }, 'processor-header-percentage-icon-alert-threshold', percentageSection, {min: 0, max: 100, digits: 0, step: 1, page: 10}, true, 0);
         
         const graphSection = this.addExpanderRow(_('History Graph'), group);
         this.addSwitchRow(this.tab + _('Show History Graph'), 'processor-header-graph', graphSection);
@@ -333,6 +338,7 @@ export default class AstraMonitorPrefs extends ExtensionPreferences {
             subtitle:  this.tab + _('Set icon name (ie: \'memory-symbolic\')') + '\n' + this.tab + _('Set to empty to disable icon override'),
         }, 'memory-header-icon-custom', iconSection, '');
         this.addColorRow({ title: this.tab + _('Icon Color') }, 'memory-header-icon-color', iconSection, '');
+        this.addColorRow({ title: this.tab + _('Icon Alert Color') }, 'memory-header-icon-alert-color', iconSection, 'rgba(235, 64, 52, 255)');
         this.addSpinRow({
             title: this.tab + _('Icon Size'),
             subtitle:  this.tab + _('Experimental feature: may require to disable/enable the extension.'),
@@ -344,6 +350,10 @@ export default class AstraMonitorPrefs extends ExtensionPreferences {
         
         const percentageSection = this.addExpanderRow(_('Usage Percentage'), group);
         this.addSwitchRow(this.tab + _('Show Usage Percentage'), 'memory-header-percentage', percentageSection);
+        this.addSpinRow({
+            title: this.tab + _('Icon Alert'),
+            subtitle:  this.tab + _('Set 0 to disable. Value is % of ram usage.'),
+        }, 'memory-header-percentage-icon-alert-threshold', percentageSection, {min: 0, max: 100, digits: 0, step: 1, page: 10}, true, 0);
         
         const valueSection = this.addExpanderRow(_('Usage Value'), group);
         this.addSwitchRow(this.tab + _('Show Usage Value'), 'memory-header-value', valueSection);
@@ -436,6 +446,7 @@ export default class AstraMonitorPrefs extends ExtensionPreferences {
             subtitle:  this.tab + _('Set icon name (ie: \'drive-harddisk-symbolic\')') + '\n' + this.tab + _('Set to empty to disable icon override'),
         }, 'storage-header-icon-custom', iconSection, '');
         this.addColorRow({ title: this.tab + _('Icon Color') }, 'storage-header-icon-color', iconSection, '');
+        this.addColorRow({ title: this.tab + _('Icon Alert Color') }, 'storage-header-icon-alert-color', iconSection, 'rgba(235, 64, 52, 255)');
         this.addSpinRow({
             title: this.tab + _('Icon Size'),
             subtitle:  this.tab + _('Experimental feature: may require to disable/enable the extension.'),
@@ -448,6 +459,10 @@ export default class AstraMonitorPrefs extends ExtensionPreferences {
         const barsSection = this.addExpanderRow(_('Main Disk'), group);
         this.addSwitchRow(this.tab + _('Show Storage Usage Bar'), 'storage-header-bars', barsSection);
         this.addSwitchRow(this.tab + _('Show Storage Usage Percentage'), 'storage-header-percentage', barsSection);
+        this.addSpinRow({
+            title: this.tab + _('Icon Alert'),
+            subtitle:  this.tab + _('Set 0 to disable. Value is % of disk usage.'),
+        }, 'storage-header-percentage-icon-alert-threshold', barsSection, {min: 0, max: 100, digits: 0, step: 1, page: 10}, true, 0);
         
         const ioSection = this.addExpanderRow(_('IO'), group);
         this.addSwitchRow(this.tab + _('Show Realtime IO Bar'), 'storage-header-io-bars', ioSection);
@@ -560,6 +575,7 @@ export default class AstraMonitorPrefs extends ExtensionPreferences {
             subtitle:  this.tab + _('Set icon name (ie: \'network-wired-symbolic\')') + '\n' + this.tab + _('Set to empty to disable icon override'),
         }, 'network-header-icon-custom', iconSection, '');
         this.addColorRow({ title: this.tab + _('Icon Color') }, 'network-header-icon-color', iconSection, '');
+        this.addColorRow({ title: this.tab + _('Icon Alert Color') }, 'network-header-icon-alert-color', iconSection, 'rgba(235, 64, 52, 255)');
         this.addSpinRow({
             title: this.tab + _('Icon Size'),
             subtitle:  this.tab + _('Experimental feature: may require to disable/enable the extension.'),
@@ -609,6 +625,7 @@ export default class AstraMonitorPrefs extends ExtensionPreferences {
             subtitle:  this.tab + _('Set icon name (ie: \'temperature-symbolic\')') + '\n' + this.tab + _('Set to empty to disable icon override'),
         }, 'sensors-header-icon-custom', iconSection, '');
         this.addColorRow({ title: this.tab + _('Icon Color') }, 'sensors-header-icon-color', iconSection, '');
+        this.addColorRow({ title: this.tab + _('Icon Alert Color') }, 'sensors-header-icon-alert-color', iconSection, 'rgba(235, 64, 52, 255)');
         this.addSpinRow({
             title: this.tab + _('Icon Size'),
             subtitle:  this.tab + _('Experimental feature: may require to disable/enable the extension.'),
@@ -857,7 +874,6 @@ export default class AstraMonitorPrefs extends ExtensionPreferences {
                 rgba.parse(reset);
                 button.set_rgba(rgba);
             });
-            row.add_suffix(resetButton);
         }
         
         row.add_suffix(button);
@@ -959,7 +975,7 @@ export default class AstraMonitorPrefs extends ExtensionPreferences {
         
         const row = new Adw.ActionRow(props);    
         
-        let spinButton = new Gtk.SpinButton({
+        const spinButton = new Gtk.SpinButton({
             halign: Gtk.Align.END,
             valign: Gtk.Align.CENTER,
             hexpand: false,
@@ -1050,7 +1066,6 @@ export default class AstraMonitorPrefs extends ExtensionPreferences {
     }
     
     /**
-     * 
      * @param {string} config 
      * @param {*} group 
      */